+++ assets/javascripts/mermaid_load.js
@@ -1,0 +1,3 @@
+$(function() {
+  mermaid.initialize({ startOnLoad: true });
+});