+++ lib/additionals/helpers.rb
@@ -285,6 +285,11 @@ module Additionals
         additionals_include_js('nv.d3.min')
     end
 
+    def additionals_load_mermaid
+      additionals_include_js('mermaid.min') +
+        additionals_include_js('mermaid_load')
+    end
+
     def additionals_load_d3plus
       additionals_include_js('d3.min') +
         additionals_include_js('d3plus.full.min')